Move the api-availability monitor to a separate package
If the monitor is in a separate package we can run the unit tests in Travis
easily.
