+++ platform-tests/src/platform/availability/monitor/monitor_suite_test.go
@@ -1,0 +1,13 @@
+package monitor_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestMonitor(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Monitor Suite")
+}
